refactoring stage 2

+++ lib/movierenderer.rb
@@ -15,15 +15,17 @@ class MovieRenderer
     output = Haml::Engine.new(template).render(@render_object)
     File.open(output_file, 'w') { |file| file.write(output) }
   end
-  
+
   def download_tmdb_posters
     progressbar = ProgressBar.create(title: "Download images from #{TMDB_IMAGE_URL}",
-                                        starting_at: 0, total: @render_object.all.count)
-    @render_object.all.each do | movie | 
+                                     starting_at: 0, total: @render_object.all.count)
+    @render_object.all.each do |movie|
       path_to_save = "#{LOCAL_IMAGES}#{movie.poster_id}"
-      File.open(path_to_save, 'wb') do |file|
-        file.write open("#{TMDB_IMAGE_URL}#{movie.poster_id}").read
-      end unless File.exist?(path_to_save)
+      unless File.exist?(path_to_save)
+        File.open(path_to_save, 'wb') do |file|
+          file.write open("#{TMDB_IMAGE_URL}#{movie.poster_id}").read
+        end
+      end
       progressbar.increment
     end
     self
